requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
pillow
requests
av

